requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
Django==1.6.2
django-mongodb-engine==0.6.0
django-tastypie==0.11.1
djangotoolbox==1.8.0
enum34==1.1.6
idna==2.8
ipaddress==1.0.22
pycparser==2.19
pymongo==3.4.0
pyOpenSSL==18.0.0
python-dateutil==2.7.5
python-mimeparse==1.6.0
pytz==2018.7
requests==2.21.0
six==1.12.0
urllib3==1.24.1
